backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
msal==1.28.0
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
httpx==0.27.0
aioapns==3.0
cryptography==42.0.7
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.3.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
